test: remove custom `newFromString` in favor of `decimal.RequireFromString` (#145)

+++ pkg/precision/precision.go
@@ -102,7 +102,7 @@ func Validate(input decimal.Decimal, maxPrecision int32) error {
 		return errors.New("input must not be negative")
 	}
 
-	precision := int32(input.Exponent())
+	precision := input.Exponent()
 	if precision < 0 {
 		precision = -precision
 	}
